ディレクトリ構成を変更し、テストプロジェクトを追加
