+++ app/Models/TimeStampModel.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace timestamp_to_slack.Models
+{
+    internal class TimeStampModel
+    {
+        public DateTime Today { get; set; }
+        public DateTime Time { get; set; }
+    }
+}
