+++ app/Models/UserModel.cs
@@ -1,0 +1,15 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace timestamp_to_slack.Models
+{
+    public class UserModel
+    {
+        public int Index { get; set; }
+        public int Id { get; set; }
+        public string Name { get; set; }
+    }
+}
