+++ app/Program.cs
@@ -6,35 +6,8 @@ using System.Text;
 using Newtonsoft;
 using Newtonsoft.Json;
 
-namespace TimeStamptoSlack
+namespace TimeStamper
 {
-    public class TimeStamp
-    {
-        public int Index { get; set; } = 0;
-        public string UserName { get; set; } = string.Empty;
-        public string Today { get; set; } = DateTime.Today.ToString("yyyy/MM/dd");
-        public string StartTime { get; set; } = string.Empty;
-        public string EndTime { get; set; } = string.Empty;
-    }
-
-    [JsonObject]
-    public class Config
-    {
-        [JsonProperty("user_Name")]
-        public string? UserName { get; set; } = string.Empty;
-        [JsonProperty("time_Stamp_File_Path")]
-        public string? TimeStampFilePath { get; set; }
-        [JsonProperty("api_Url")]
-        public string? ApiUrl { get; set; }
-    }
-
-    [JsonObject]
-    public class Slack
-    {
-        [JsonProperty("text")]
-        public string Text { get; set; } = string.Empty;
-    }
-
     class Program
     {
         /// <summary>
@@ -66,68 +39,6 @@ namespace TimeStamptoSlack
             
             Console.WriteLine(WriteCSV(isStart, config.UserName, config.TimeStampFilePath));
             Console.ReadKey();
-        }
-
-        /// <summary>
-        /// config.jsonから打刻ファイルのパスを取得する。
-        /// </summary>
-        /// <returns></returns>
-        public static Config GetConfig()
-        {
-            var configFilePath = "./config.json";
-            Config config = new Config()
-            {
-                TimeStampFilePath = "./Dakoku.csv",
-            };
-
-            if (!File.Exists(configFilePath))
-            {
-                Console.WriteLine("打刻ファイルを作成します。");
-                Console.WriteLine("ユーザー名を入力してください。");
-                config.UserName = Console.ReadLine();
-                Console.WriteLine("Slack Webhool URLを入力してください。");
-                config.ApiUrl = Console.ReadLine();
-
-                using (StreamWriter file = File.CreateText(configFilePath))
-                {
-                    JsonSerializer serializer = new JsonSerializer();
-                    serializer.Serialize(file, config);
-                }
-            }
-            else
-            {
-                var jsonString = File.ReadAllText(configFilePath);
-                var configJson = JsonConvert.DeserializeObject<Config>(jsonString);
-
-                if (configJson != null)
-                {
-                    config = configJson;
-                }
-            }
-
-            Console.WriteLine("File path: " + config.TimeStampFilePath);
-            return config;
-        }
-
-        /// <summary>
-        /// SlackのIn coming webhookを使用し、任意のチャンネルに、始業、終業の投稿を行う。
-        /// </summary>
-        /// <returns></returns>
-        static async void PostSlack(string webhook, string postText)
-        {
-            Slack slack = new Slack()
-            {
-                Text = postText
-            };
-
-            var json = JsonConvert.SerializeObject(slack);
-
-            using (var client = new HttpClient())
-            {
-                var content = new StringContent(json, Encoding.UTF8, "application/json");
-
-                var response = await client.PostAsync(webhook, content);
-            };
         }
 
         /// <summary>
