+++ test/TimeStamper.Test/TimeStamperAppTest.cs
@@ -1,0 +1,11 @@
+namespace TimeStamper.Test
+{
+    public class TimeStamperAppTest
+    {
+        [Fact]
+        public void CorrectCalculationElapsedTime()
+        {
+            
+        }
+    }
+}
